+++ src/client/main.jsx
@@ -14,9 +14,10 @@ import '@fontsource/roboto/300.css';
 import '@fontsource/roboto/400.css';
 import '@fontsource/roboto/500.css';
 import '@fontsource/roboto/700.css';
+import { App } from './App.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
     <StrictMode>
-        <div>agenda-display</div>
+        <App />
     </StrictMode>
 );